Remove dtype parameter from tokenizer input tensor conversion to fix device placement compatibility issues
- Change `.to(self._hf_model.device, dtype=self._hf_model.dtype)` to `.to(self._hf_model.device)` in HuggingFaceClient.chat() to avoid dtype mismatch errors when moving tokenized inputs to model device
- Allows PyTorch to handle dtype conversion automatically during model forward pass

+++ modules/llm_client.py
@@ -327,7 +327,7 @@ class HuggingFaceClient:
                 tokenize=True,
                 return_dict=True,
                 return_tensors="pt",
-            ).to(self._hf_model.device, dtype=self._hf_model.dtype)
+            ).to(self._hf_model.device)
 
             input_len = inputs["input_ids"].shape[-1]
